+++ src/main/java/com/ecommerce/project/service/ProductService.java
@@ -7,18 +7,18 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
 
 public interface ProductService {
-    ProductDTO addProduct(Long categoryId, ProductDTO productDTO);
+    ProductDTO addProduct(Integer categoryId, ProductDTO productDTO);
 
     ProductResponse getAllProducts(Integer pageNumber, Integer pageSize, String sortBy, String sortOrder);
 
-    ProductResponse getProductsByCategory(Integer pageNumber, Integer pageSize, String sortBy, String sortOrder, Long categoryId);
+    ProductResponse getProductsByCategory(Integer pageNumber, Integer pageSize, String sortBy, String sortOrder, Integer categoryId);
 
     ProductResponse getProductsByKeyword(Integer pageNumber, Integer pageSize, String sortBy, String sortOrder, String keyword);
 
-    ProductDTO updateProduct(ProductDTO productDTO, Long productId);
+    ProductDTO updateProduct(ProductDTO productDTO, Integer productId);
 
-    String deleteProduct(Long productId);
+    String deleteProduct(Integer productId);
 
-    ProductDTO updateProductImage(MultipartFile image, Long productId) throws IOException;
+    ProductDTO updateProductImage(MultipartFile image, Integer productId) throws IOException;
 
 }
